moved searchbar function to app.js, added use effect but ran into errors.
