+++ src/App.js
@@ -1,16 +1,16 @@
 import "./App.css";
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { HeaderComponent } from "./components/Header/HeaderComponent";
 import { SidebarComponent } from "./components/Sidebar/SidebarComponent";
 import { FeedComponent } from "./components/Feed/FeedComponent";
 import { FooterComponent } from "./components/Footer/FooterComponent.jsx";
 import { CharacterCard } from "./components/STRETCH_Cards/CharacterCard";
 
-// const api = {
-//   books: "https://www.anapioficeandfire.com/api/books",
-//   houses: "https://www.anapioficeandfire.com/api/houses",
-//   characters: "https://www.anapioficeandfire.com/api/characters",
-// };
+const api = {
+  books: "https://www.anapioficeandfire.com/api/books",
+  houses: "https://www.anapioficeandfire.com/api/houses",
+  characters: "https://www.anapioficeandfire.com/api/characters",
+};
 
 const App = () => {
   // TODO - this is the "main" component for our app, and it will include all the global state that we care about
@@ -51,12 +51,29 @@ const App = () => {
   const [feed, setFeed] = useState([]);
   const [query, setQuery] = useState("");
 
+  // fetch function for entering a name
+  //moved from headercomponent 
+   const search = (query) => {
+    return fetch(`${api.characters}?name=${query}`)
+      .then((resp) => {
+      return resp.json().then((data) => {
+        setQuery("");
+        console.log(data);
+       });
+      })
+  };
+
+  useEffect(() => {
+    search(query).then((results) => setFeed(results));
+  },[query]);
+  
+
   return (
-    //gotta return the functions here somehow...
     <div className="app">
       <HeaderComponent query={query} setQuery={setQuery} />
-
+     
       <SidebarComponent query={query} setQuery={setQuery} />
+
       {/* <button
         className="fetchButton"
         onClick={() => {
@@ -70,6 +87,6 @@ const App = () => {
       <FooterComponent />
     </div>
   );
-};
+    };
 
 export default App;
